requirements.txt
fastapi==0.68.3
uvicorn==0.15.0
pydantic==1.9.0
confluent-kafka==1.8.0
python-dotenv==0.19.2
redis==4.0.3
